refactor : 비동기 처리 로직 리팩토링
